// frontend/src/types/tourism/payment/v1/payment.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.7
//   protoc               unknown
// source: tourism/payment/v1/payment.proto

/* eslint-disable */

export const protobufPackage = "tourism.payment.v1";

export interface OrderItem {
  id: string;
  tourId: string;
  tourName: string;
  price: number;
  addedAt: string;
}

export interface ShoppingCart {
  id: string;
  userId: string;
  items: OrderItem[];
  totalPrice: number;
}

export interface TourPurchaseToken {
  id: string;
  userId: string;
  tourId: string;
  tourName: string;
  price: number;
  purchasedAt: string;
}

export interface GetCartRequest {
}

export interface AddToCartRequest {
  tourId: string;
}

export interface RemoveFromCartRequest {
  tourId: string;
}

export interface CheckoutRequest {
}

export interface CheckoutResponse {
  tokens: TourPurchaseToken[];
}

export interface ListMyTokensRequest {
}

export interface ListMyTokensResponse {
  tokens: TourPurchaseToken[];
}

export interface HasTokenRequest {
  tourId: string;
}

export interface HasTokenResponse {
  hasToken: boolean;
}